Add error messages on backend responses
Make the navigation bar and tab bar background color and buttons color consistent throu the entire app

Add Select image on sign up screen so the user can select a profile picture.
Upload the picture when the create button is pressed.
Add that image url to the sign up request.